feat: 웹소켓 커스텀 훅 제작, UI 컴포넌트 분리
