+++ src/hooks/chat/useCreateChatRoom.ts
@@ -9,7 +9,7 @@ export const useCreateChatRoom = () => {
     mutationFn: createChatRoom,
     onSuccess: ({ chatRoomId }) => {
       queryClient.invalidateQueries({
-        queryKey: CHAT_KEYS.CHAT_LIST,
+        queryKey: [CHAT_KEYS.CHAT_LIST],
       });
       alert(`${chatRoomId}번 채팅방 생성완료!`);
     },
